Wrote documentation for Game, fixed a bug where Game.checkAvailableMoves didn't check both players, rewrote checkVictory to print the final scores

+++ src/pij/main/Player.java
@@ -21,7 +21,7 @@ public abstract class Player {
 	private final List<CharacterTile> tileRack = new ArrayList<>();
 
 	/** A player's score. */
-	private double score = 0;
+	private int score = 0;
 
 	private final Board board;
 
@@ -45,7 +45,7 @@ public abstract class Player {
 	 * 
 	 * @return the player's score.
 	 */
-	public double getScore() {
+	public int getScore() {
 		return this.score;
 	}
